refactor: separate client logic from pages
